Add MapBox in library
